Add add tasks API to DB layer
