+++ common/persistence/cassandra/cassandraPersistence.go
@@ -1877,6 +1877,67 @@ func (d *cassandraPersistence) ListConcreteExecutions(
 	return response, nil
 }
 
+func (d *cassandraPersistence) AddTasks(request *p.AddTasksRequest) error {
+	batch := d.session.NewBatch(gocql.LoggedBatch)
+
+	if err := applyTasks(
+		batch,
+		d.shardID,
+		request.NamespaceID,
+		request.WorkflowID,
+		request.RunID,
+		request.TransferTasks,
+		request.TimerTasks,
+		request.ReplicationTasks,
+		request.VisibilityTasks,
+	); err != nil {
+		return err
+	}
+
+	batch.Query(templateUpdateLeaseQuery,
+		request.RangeID,
+		d.shardID,
+		rowTypeShard,
+		rowTypeShardNamespaceID,
+		rowTypeShardWorkflowID,
+		rowTypeShardRunID,
+		defaultVisibilityTimestamp,
+		rowTypeShardTaskID,
+		request.RangeID,
+	)
+
+	previous := make(map[string]interface{})
+	applied, iter, err := d.session.MapExecuteBatchCAS(batch, previous)
+	defer func() {
+		if iter != nil {
+			_ = iter.Close()
+		}
+	}()
+
+	if err != nil {
+		if isTimeoutError(err) {
+			// Write may have succeeded, but we don't know
+			// return this info to the caller so they have the option of trying to find out by executing a read
+			return &p.TimeoutError{Msg: fmt.Sprintf("AddTasks timed out. Error: %v", err)}
+		} else if isThrottlingError(err) {
+			return serviceerror.NewResourceExhausted(fmt.Sprintf("AddTasks operation failed. Error: %v", err))
+		}
+		return serviceerror.NewInternal(fmt.Sprintf("AddTasks operation failed. Error: %v", err))
+	}
+	if !applied {
+		if previousRangeID, ok := previous["range_id"].(int64); ok && previousRangeID != request.RangeID {
+			// CreateWorkflowExecution failed because rangeID was modified
+			return &p.ShardOwnershipLostError{
+				ShardID: d.shardID,
+				Msg:     fmt.Sprintf("Failed to add tasks.  Request RangeID: %v, Actual RangeID: %v", request.RangeID, previousRangeID),
+			}
+		} else {
+			return serviceerror.NewInternal("AddTasks operation failed: %v")
+		}
+	}
+	return nil
+}
+
 func (d *cassandraPersistence) GetTransferTask(request *p.GetTransferTaskRequest) (*p.GetTransferTaskResponse, error) {
 	shardID := d.shardID
 	taskID := request.TaskID
